usando node-fetch para mejor compatibilidad

+++ commands/card.js
@@ -9,12 +9,16 @@ const getCard = async (message) => {
         username: message.author.username,
     }
     const user = await userService.getUser(auxUser)
+    const card = await cardService.getCard()
     if (user.balance < 100) {
         message.reply(createEmbedText( 0xFF0000, "You do not have enough money to do this!\nYou need 100 coins to buy a card!(Your balance is: " + user.balance + ")"))
         return
     }
+    if(!card.length){
+        message.reply(createEmbedText( 0xFF0000, "No cards available!"))
+        return
+    }
     await userService.removeBalance(userId, 100)
-    const card = await cardService.getCard()
     const embed = await createEmbedCard( 0x00569D, card[0])
     await userService.addCard(userId, card[0]._id)
     message.reply(embed)
